smart - added component
